+++ fyne/fyne.go
@@ -5,13 +5,12 @@ import (
 
 	"fyne.io/fyne/app"
 	godcrApp "github.com/raedahgroup/godcr/app"
-	"github.com/raedahgroup/godcr/fyne/pages"
 )
 
 func LaunchFyne(ctx context.Context, walletMiddleware godcrApp.WalletMiddleware) {
 	fyneApp := app.New()
 	window := fyneApp.NewWindow(godcrApp.DisplayName)
-	window.SetContent(pages.Menu(pages.ShowSyncWindow(walletMiddleware, window, fyneApp), window, fyneApp))
+	window.SetContent(ShowSyncWindow(walletMiddleware, window, fyneApp))
 	window.ShowAndRun()
 
 }
